requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
requests==2.32.5
python-dotenv==1.2.1
pytz==2025.2
aiohttp==3.9.5
pandas==2.2.2
pykrx==1.0.51
finance-datareader-0.9.101
torch
